refactor: update booking status page with dynamic component mapping and data hook
